Remove FormatterUSTest.scala, extract Locale constructor tests from it to seperate suite
